+++ public/JS/inventarioFuncoes.js
@@ -1,0 +1,1 @@
+//funções da página de inventário.html e consultarEditarItensInventario.html
