+++ views/cadastrarUsuarioInternal.ejs
@@ -1,0 +1,172 @@
+<!DOCTYPE html>
+<html lang="en" dir="ltr">
+
+<head>
+  <meta charset="utf-8">
+  <title>Cadastrar usuários</title>
+  <!-- fontes do google -->
+  <link rel="preconnect" href="https://fonts.googleapis.com">
+  <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
+  <link href="https://fonts.googleapis.com/css2?family=Montserrat:wght@100&display=swap" rel="stylesheet">
+  <!-- css e js links-->
+  <script src="https://kit.fontawesome.com/cb4069a99b.js" crossorigin="anonymous"></script>
+  <link rel="stylesheet" href="css/styles.css">
+  <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.0/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-KyZXEAg3QhqLMpG8r+8fhAXLRk2vvoC2f3B09zVXn8CA5QIVfZOJ3BCsw2P0p/We" crossorigin="anonymous">
+  <link rel="icon" href="images\favicon.ico">
+  <script src="JS/functions.js" type="text/javascript"></script>
+</head>
+
+<body class="paginaInicial">
+  <!-- Seção cabeçalho que é utilizada em todas as telas internas do site-->
+  <div class=containerTela>
+    <!--barra de menu superior></!-->
+    <div class="topnav">
+      <span class="bemVindo">Cadastrar usuários</span>
+      <a href="paginaPrincipal.html"> <i class="fas fa-house-user"> </i> Página Inicial</a>
+      <div class="topnav-right">
+        <a href="index.html"><i class="fas fa-sign-out-alt"></i> Sair</a>
+      </div>
+    </div>
+
+    <!--barra de menu lateral></!-->
+
+    <div class="barraLateral">
+      <div class="vertical-menu">
+        <br>
+        <a href="#">Relatórios</a>
+        <button class="dropdown-btn">Ordens de serviço
+          <i class="fa fa-caret-down"></i>
+        </button>
+        <div class="dropdown-container">
+          <a href="#">Criar ordens de serviço</a>
+          <a href="#">Consultar e editar ordens de serviço</a>
+        </div>
+        <button class="dropdown-btn">Ordens preventivas
+          <i class="fa fa-caret-down"></i>
+        </button>
+        <div class="dropdown-container">
+          <a href="#">Criar ordem preventiva</a>
+          <a href="#">Consultar e editar ordens preventivas</a>
+        </div>
+        <button class="dropdown-btn">Máquinas
+          <i class="fa fa-caret-down"></i>
+        </button>
+        <div class="dropdown-container">
+          <a href="#">Cadastrar máquinas</a>
+          <a href="#">Consultar e editar máquinas</a>
+          <a href="#">Excluir máquinas</a>
+          <a href="#">Dados de processo</a>
+        </div>
+        <button class="dropdown-btn">Usuários
+          <i class="fa fa-caret-down"></i>
+        </button>
+        <div class="dropdown-container">
+          <a href="cadastrarUsuarioInternal.html">Cadastrar usuários</a>
+          <a href="consultarEditarUsuarios.html">Consultar e editar usuários</a>
+        </div>
+        <button class="dropdown-btn">Inventário
+          <i class="fa fa-caret-down"></i>
+        </button>
+        <div class="dropdown-container">
+          <a href="inventario.html">Cadastrar Itens</a>
+          <a href="consultarEditarItensInventario.html">Consultar e editar Itens</a>
+        </div>
+        <a href="#">Contate-nos!</a>
+      </div>
+    </div>
+
+    <script>
+      /* função para ativar os botões da barra lateral, dropdownbtn*/
+      var dropdown = document.getElementsByClassName("dropdown-btn");
+      var i;
+
+      for (i = 0; i < dropdown.length; i++) {
+        dropdown[i].addEventListener("click", function() {
+          this.classList.toggle("active");
+          var dropdownContent = this.nextElementSibling;
+          if (dropdownContent.style.display === "block") {
+            dropdownContent.style.display = "none";
+          } else {
+            dropdownContent.style.display = "block";
+          }
+        });
+      }
+    </script>
+
+    <!--Conteúdo interno da página de criação do usuário></!-->
+    <div class="conteudo">
+      <div class="conteudoEsquerdoUsuario">
+
+        <form>
+          <div class="mb-3">
+            <label for="nomeUsuario" class="form-label">Nome de usuário</label>
+            <input type="text" class="form-control" id="nomeUsuario" maxlength="50">
+            <div id="nomeUsuarioHelp" class="form-text">O nome de usuário deve ter pelo menos 6 caracteres e não deve possuir " ".</div>
+          </div>
+
+          <div class="mb-3">
+            <label for="nomeCompletoUsuario" class="form-label">Nome completo</label>
+            <input type="text" class="form-control" id="nomeUsuario" aria-describedby="nomeHelp">
+            <div id="nomeHelp" class="form-text">Por favor digite seu nome completo.</div>
+          </div>
+
+          <div class="mb-3">
+            <label for="usuarioEmail1" class="form-label">Endereço de E-mail</label>
+            <input type="email" class="form-control" id="usuarioEmail1" maxlength="100">
+          </div>
+
+          <div class="mb-3">
+            <label for="usuarioPassword1" class="form-label">Senha</label>
+            <input type="password" class="form-control" id="usuarioPassword1">
+          </div>
+          <div class="mb-3">
+            <label for="usuarioPassword2" class="form-label">Confirmação de senha</label>
+            <input type="password" class="form-control" id="usuarioPassword2">
+          </div>
+          <button type="submit" class="btn btn-primary"><i class="fas fa-save"></i> Salvar</button>
+        </form>
+
+      </div>
+      <div class="conteudoDireitoUsuario">
+
+        <div class="dataNascimentoTipoUsuario">
+          <div class="mb-3">
+          <label class="form-label"> Data de nascimento:</label>
+          <input class="form-control" type="date" name="" value="" id="dataNascimento"></input>
+          <br>
+          </div>
+
+
+          <div class="mb-3">
+          <label class ="form-label">Tipo de usuário: </label>
+          </div>
+          <div class="form-check">
+          <input class="form-check-input" type="radio" name="flexRadioDefault" id="tipoAdministrador">
+          <label class="form-check-label" for="tipoAdministrador">
+            Administrador
+          </label>
+        </div>
+        <br>
+        <div class="form-check">
+          <input class="form-check-input" type="radio" name="flexRadioDefault" id="tipoTecnico" checked>
+          <label class="form-check-label" for="tipoTecnico">
+            Técnico
+          </label>
+          </div>
+          <br>
+          <div class="form-check">
+            <input class="form-check-input" type="radio" name="flexRadioDefault" id="tipoUsuario" checked>
+            <label class="form-check-label" for="tipoUsuario">
+              Usuário
+            </label>
+        </div>
+        </div>
+      </div>
+    </div>
+
+  </div>
+
+<script src="JS/usuariosFuncoes.js" type="text/javascript"></script>
+</body>
+
+</html>
