+++ views/consultarEditarItensInventario.ejs
@@ -1,0 +1,186 @@
+<!DOCTYPE html>
+<html lang="en" dir="ltr">
+
+<head>
+  <meta charset="utf-8">
+  <title>Consultar e editar itens</title>
+  <!-- fontes do google -->
+  <link rel="preconnect" href="https://fonts.googleapis.com">
+  <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
+  <link href="https://fonts.googleapis.com/css2?family=Montserrat:wght@100&display=swap" rel="stylesheet">
+  <!-- css e js links-->
+  <script src="https://kit.fontawesome.com/cb4069a99b.js" crossorigin="anonymous"></script>
+  <link rel="stylesheet" href="css/styles.css">
+  <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.0/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-KyZXEAg3QhqLMpG8r+8fhAXLRk2vvoC2f3B09zVXn8CA5QIVfZOJ3BCsw2P0p/We" crossorigin="anonymous">
+  <link rel="icon" href="images\favicon.ico">
+
+</head>
+
+<body class="paginaInicial">
+  <!-- Seção cabeçalho que é utilizada em todas as telas internas do site-->
+  <div class=containerTela>
+    <!--barra de menu superior></!-->
+    <div class="topnav">
+      <span class="bemVindo">Inventário - Consultar e editar itens</span>
+      <a href="paginaPrincipal.html"> <i class="fas fa-house-user"> </i> Página Inicial</a>
+      <div class="topnav-right">
+        <a href="index.html"><i class="fas fa-sign-out-alt"></i> Sair</a>
+      </div>
+    </div>
+
+    <!--barra de menu lateral></!-->
+
+
+    <div class="barraLateral">
+      <div class="vertical-menu">
+        <br>
+        <a href="#">Relatórios</a>
+        <button class="dropdown-btn">Ordens de serviço
+          <i class="fa fa-caret-down"></i>
+        </button>
+        <div class="dropdown-container">
+          <a href="#">Criar ordens de serviço</a>
+          <a href="#">Consultar e editar ordens de serviço</a>
+        </div>
+        <button class="dropdown-btn">Ordens preventivas
+          <i class="fa fa-caret-down"></i>
+        </button>
+        <div class="dropdown-container">
+          <a href="#">Criar ordem preventiva</a>
+          <a href="#">Consultar e editar ordens preventivas</a>
+        </div>
+        <button class="dropdown-btn">Máquinas
+          <i class="fa fa-caret-down"></i>
+        </button>
+        <div class="dropdown-container">
+          <a href="#">Cadastrar máquinas</a>
+          <a href="#">Consultar e editar máquinas</a>
+          <a href="#">Excluir máquinas</a>
+          <a href="#">Dados de processo</a>
+        </div>
+        <button class="dropdown-btn">Usuários
+          <i class="fa fa-caret-down"></i>
+        </button>
+        <div class="dropdown-container">
+          <a href="cadastrarUsuarioInternal.html">Cadastrar usuários</a>
+          <a href="consultarEditarUsuarios.html">Consultar e editar usuários</a>
+        </div>
+        <button class="dropdown-btn">Inventário
+          <i class="fa fa-caret-down"></i>
+        </button>
+        <div class="dropdown-container">
+          <a href="inventario.html">Cadastrar Itens</a>
+          <a href="consultarEditarItensInventario.html">Consultar e editar Itens</a>
+        </div>
+        <a href="#">Contate-nos!</a>
+      </div>
+    </div>
+
+    <script>
+      /* função para ativar os botões da barra lateral, dropdownbtn*/
+      var dropdown = document.getElementsByClassName("dropdown-btn");
+      var i;
+
+      for (i = 0; i < dropdown.length; i++) {
+        dropdown[i].addEventListener("click", function() {
+          this.classList.toggle("active");
+          var dropdownContent = this.nextElementSibling;
+          if (dropdownContent.style.display === "block") {
+            dropdownContent.style.display = "none";
+          } else {
+            dropdownContent.style.display = "block";
+          }
+        });
+      }
+    </script>
+
+
+    <div class="conteudo">
+      <div class="conteudoEsquerdoInventario">
+
+        <form>
+          <div class="mb-3">
+            <label for="idItem" class="form-label">ID Item</label>
+            <input type="text" class="form-control" id="idItem" maxlength="50">
+            <div id="IdItemHelp" class="form-text">Insira o ID do item</div>
+          </div>
+
+          <div class="mb-3">
+            <div class="form-group">
+              <label for="descricaoItem" class="form-label">Descrição do item</label>
+              <textarea class="form-control" id="descricaoItem" rows="3" maxlength="100"></textarea>
+              <div id="descricaoitemHelp" class="form-text">Insira uma breve descrição do item cadastrado.</div>
+            </div>
+          </div>
+
+          <div class="mb-3">
+            <label for="maquinaItem" class="form-label">Máquina</label>
+            <input type="text" class="form-control" id="maquinaItem" aria-describedby="nomeHelp">
+            <div id="maquinaItemHelp" class="form-text">Insira a identificação da máquina em que o item é utilizado.</div>
+          </div>
+
+          <div class="mb-3">
+            <label for="vendedorItem" class="form-label">Vendedor</label>
+            <input type="email" class="form-control" id="vendedorItem" maxlength="100">
+            <div id="vendedorItemHelp" class="form-text">Insira o nome da empresa ou do vendedor do item.</div>
+          </div>
+          <button type="button" class="btn btn-primary"><i class="fas fa-search"></i> Procurar</button>
+          <button type="button" class="btn btn-primary"><i class="fas fa-eraser"></i> Cancelar</button>
+          <a href="paginaPrincipal.html" class="btn btn-primary"><i class="fas fa-long-arrow-alt-left"></i> Voltar</a>
+        </form>
+
+      </div>
+      <div class="conteudoDireitoInventario">
+
+        <form>
+          <div class="mb-3">
+            <label for="marcaItem" class="form-label">Marca</label>
+            <input type="text" class="form-control" id="marcaItem">
+            <div id="marcaItemHelp" class="form-text">Insira a marca do item.</div>
+          </div>
+
+          <div class="mb-3">
+            <label for="modeloItem" class="form-label">Modelo</label>
+            <input type="text" class="form-control" id="modeloItem">
+            <div id="modeloItemHelp" class="form-text">Insira o modelo do item.</div>
+          </div>
+          <br>
+          <div class="mb-3">
+            <div class="form-check">
+              <input class="form-check-input" type="checkbox" name="todosItens" id="todosItens">
+              <label class="form-check-label" for="todosItens">
+                Pesquisar todos?
+              </label>
+              <div id="todosItensHelp" class="form-text">Se marcado esse campo, a pesquisa apresentará todos itens cadastrados</div>
+            </div>
+          </div>
+
+          <div class="form-group">
+            <label for="resultadoItens" class="form-label">Resultado da pesquisa:</label>
+
+            <select multiple class="form-control" id="resultadoItens">
+              <option>1</option>
+              <option>2</option>
+              <option>3</option>
+              <option>4</option>
+              <option>5</option>
+            </select>
+            <div id="pesqueisaItensHelp" class="form-text">Selecione o item que deseja alterar ou excluir</div>
+          </div>
+          <div class="containerBtnInventarioEdit">
+            <br>
+            <button type="submit" class="btn btn-primary"><i class="fas fa-save"></i> Salvar</button>
+            <button type="button" class="btn btn-primary"><i class="fas fa-edit"></i> Editar</button>
+            <button type="button" class="btn btn-primary"><i class="fas fa-trash-alt"></i> Excluir</button>
+          </div>
+
+
+      </div>
+    </div>
+
+  </div>
+
+  <script src="JS/inventarioFuncoes.js" type="text/javascript"></script>
+</body>
+
+</html>
