+++ views/consultarEditarUsuarios.ejs
@@ -1,0 +1,184 @@
+<!DOCTYPE html>
+<html lang="en" dir="ltr">
+
+<head>
+  <meta charset="utf-8">
+  <title>Consultar e editar usuários</title>
+  <!-- fontes do google -->
+  <link rel="preconnect" href="https://fonts.googleapis.com">
+  <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
+  <link href="https://fonts.googleapis.com/css2?family=Montserrat:wght@100&display=swap" rel="stylesheet">
+  <!-- css e js links-->
+  <script src="https://kit.fontawesome.com/cb4069a99b.js" crossorigin="anonymous"></script>
+  <link rel="stylesheet" href="css/styles.css">
+  <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.0/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-KyZXEAg3QhqLMpG8r+8fhAXLRk2vvoC2f3B09zVXn8CA5QIVfZOJ3BCsw2P0p/We" crossorigin="anonymous">
+  <link rel="icon" href="images\favicon.ico">
+
+</head>
+
+<body class="paginaInicial">
+  <!-- Seção cabeçalho que é utilizada em todas as telas internas do site-->
+  <div class=containerTela>
+    <!--barra de menu superior></!-->
+    <div class="topnav">
+      <span class="bemVindo">Consultar e editar usuários</span>
+      <a href="paginaPrincipal.html"> <i class="fas fa-house-user"> </i> Página Inicial</a>
+      <div class="topnav-right">
+        <a href="index.html"><i class="fas fa-sign-out-alt"></i> Sair</a>
+      </div>
+    </div>
+
+    <!--barra de menu lateral></!-->
+
+    <div class="barraLateral">
+      <div class="vertical-menu">
+        <br>
+        <a href="#">Relatórios</a>
+        <button class="dropdown-btn">Ordens de serviço
+          <i class="fa fa-caret-down"></i>
+        </button>
+        <div class="dropdown-container">
+          <a href="#">Criar ordens de serviço</a>
+          <a href="#">Consultar e editar ordens de serviço</a>
+        </div>
+        <button class="dropdown-btn">Ordens preventivas
+          <i class="fa fa-caret-down"></i>
+        </button>
+        <div class="dropdown-container">
+          <a href="#">Criar ordem preventiva</a>
+          <a href="#">Consultar e editar ordens preventivas</a>
+        </div>
+        <button class="dropdown-btn">Máquinas
+          <i class="fa fa-caret-down"></i>
+        </button>
+        <div class="dropdown-container">
+          <a href="#">Cadastrar máquinas</a>
+          <a href="#">Consultar e editar máquinas</a>
+          <a href="#">Excluir máquinas</a>
+          <a href="#">Dados de processo</a>
+        </div>
+        <button class="dropdown-btn">Usuários
+          <i class="fa fa-caret-down"></i>
+        </button>
+        <div class="dropdown-container">
+          <a href="cadastrarUsuarioInternal.html">Cadastrar usuários</a>
+          <a href="consultarEditarUsuarios.html">Consultar e editar usuários</a>
+        </div>
+        <button class="dropdown-btn">Inventário
+          <i class="fa fa-caret-down"></i>
+        </button>
+        <div class="dropdown-container">
+          <a href="inventario.html">Cadastrar Itens</a>
+          <a href="consultarEditarItensInventario.html">Consultar e editar Itens</a>
+        </div>
+        <a href="#">Contate-nos!</a>
+      </div>
+    </div>
+
+    <script>
+      /* função para ativar os botões da barra lateral, dropdownbtn*/
+      var dropdown = document.getElementsByClassName("dropdown-btn");
+      var i;
+
+      for (i = 0; i < dropdown.length; i++) {
+        dropdown[i].addEventListener("click", function() {
+          this.classList.toggle("active");
+          var dropdownContent = this.nextElementSibling;
+          if (dropdownContent.style.display === "block") {
+            dropdownContent.style.display = "none";
+          } else {
+            dropdownContent.style.display = "block";
+          }
+        });
+      }
+    </script>
+
+
+    <div class="conteudo">
+      <div class="conteudoEsquerdoUsuario">
+
+        <p class="usuariosCabecalho">Pesquisar usuário</p>
+        <div class="input-group rounded">
+          <input type="search" class="form-control rounded" placeholder="Nome de usuário (login)" aria-label="Nome de usuário" aria-describedby="search-addon" />
+          <span class="input-group-text border-0" id="procurarNomeUsuario">
+            <i class="fas fa-search"></i>
+          </span>
+        </div>
+        <br>
+        <div class="input-group rounded">
+          <input type="search" class="form-control rounded" placeholder="Nome completo do usuário" aria-label="Nome completo do usuário" aria-describedby="search-addon" />
+          <span class="input-group-text border-0" id="procurarNomeCompleto">
+            <i class="fas fa-search"></i>
+          </span>
+        </div>
+        <br>
+        <div class="input-group rounded">
+          <input type="search" class="form-control rounded" placeholder="E-mail" aria-label="E-mail" aria-describedby="search-addon" />
+          <span class="input-group-text border-0" id="procurarEmail">
+            <i class="fas fa-search"></i>
+          </span>
+        </div>
+        <br>
+
+        <div class="mb-3">
+          <label class="form-label">Tipo de usuário: </label>
+        </div>
+        <div class="form-check">
+          <input class="form-check-input" type="checkbox" name="tipoAdministrador" id="tipoAdministrador">
+          <label class="form-check-label" for="tipoAdministrador">
+            Administrador
+          </label>
+        </div>
+        <br>
+        <div class="form-check">
+          <input class="form-check-input" type="checkbox" name="tipoTecnico" id="tipoTecnico">
+          <label class="form-check-label" for="tipoTecnico">
+            Técnico
+          </label>
+        </div>
+        <br>
+        <div class="form-check">
+          <input class="form-check-input" type="checkbox" name="tipoUsuario" id="tipoUsuario">
+          <label class="form-check-label" for="tipoUsuario">
+            Usuário
+          </label>
+        </div>
+        <br>
+        <div class="btn-usuario">
+          <button type="button" class="btn btn-primary"><i class="fas fa-search"></i> Procurar</button>
+          <button type="button" class="btn btn-primary"><i class="fas fa-eraser"></i> Cancelar</button>
+          <a href="paginaPrincipal.html" class="btn btn-primary"><i class="fas fa-long-arrow-alt-left"></i> Voltar</a>
+        </div>
+      </div>
+
+
+      <div class="conteudoDireitoUsuario">
+
+        <p class="usuariosCabecalho">Lista de usuários encontrados:</p>
+
+        <div class="resultadoUsuarios">
+          <table id="tabelaUsuariosPesquisados">
+            <tr>
+              <td class = "celulasUsuarioPesquisado">ID Usuário</td>
+              <td class = "celulasUsuarioPesquisado">Login</td>
+              <td class = "celulasNomePesquisado">Nome Completo</td>
+            </tr>
+
+          </table>
+        </div>
+
+        <div class="btn-usuarioDireito">
+          <button type="button" class="btn btn-primary"><i class="fas fa-edit"></i> Editar</button>
+          <button type="button" class="btn btn-primary"><i class="fas fa-trash-alt"></i> Excluir</button>
+        </div>
+
+
+      </div>
+    </div>
+
+  </div>
+
+  <script src="JS/usuariosFuncoes.js" type="text/javascript"></script>
+</body>
+
+</html>
