+++ views/index.ejs
@@ -18,6 +18,10 @@
 
 <body>
 
+  <form class="" action="/>" method="post">
+
+  </form>
+
   <div class="logo">
     <h1>Clever</h1>
     <h2>Maintenance Software</h2>
@@ -27,7 +31,8 @@
       <tr>
         <td><label>Login:</label></td>
 
-<form  method="post">
+
+
 
         <td><input type="text" name="nome" value="" id="login" size="40" maxlength="50"></td>
         <td><a href="/confirmarUsuario">Esqueceu usuário?</a></td>
