+++ views/paginaPrincipal.ejs
@@ -22,7 +22,7 @@
     <!--barra de menu superior></!-->
     <div class="topnav">
       <span class="bemVindo">Bem Vindo ao Clever!</span>
-      <a href="paginaPrincipal.html"> <i class="fas fa-house-user"> </i> Página Inicial</a>
+      <a href="paginaPrincipal.ejs"> <i class="fas fa-house-user"> </i> Página Inicial</a>  <!--alteração do html para ejs
       <div class="topnav-right">
         <a href="index.html"><i class="fas fa-sign-out-alt"></i> Sair</a>
       </div>
@@ -75,9 +75,6 @@
     </div>
 
   </div>
-
-
-
 
 </body>
